config: fix compatibility issue
It seems that the configuration name has changed and this break the
compatibility:

Before (v1.22 or before):

    <org.jenkinsci.plugins.cppcheck.CppcheckPublisher>
      <cppcheckConfig>
        <pattern>**/cppcheck.xml</pattern>

After (v1.23):

    <org.jenkinsci.plugins.cppcheck.CppcheckPublisher plugin="cppcheck@1.23">
      <config> <!-- Not cppcheckConfig anymore -->
        <pattern>**/cppcheck.xml</pattern>

You can check the generated config.xml file by looking at this address:

    https://<JENKINS_URL>/job/<JOBNAME>/config.xml

I am not sure how the XML scheme is generated but I guess it inspects
the name of the variables. Then this change should fix the
compatibility issue.

Issue: JENKINS-52854
Fixes: 32710c4 (Fixed storing plugin-settings)
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>